+++ internal/cli/gitalybackup/restore.go
@@ -35,8 +35,8 @@ type restoreSubcommand struct {
 
 func (cmd *restoreSubcommand) flags(ctx *cli.Command) {
 	cmd.backupPath = ctx.String("path")
-	cmd.parallel = int(ctx.Int("parallel"))
-	cmd.parallelStorage = int(ctx.Int("parallel-storage"))
+	cmd.parallel = ctx.Int("parallel")
+	cmd.parallelStorage = ctx.Int("parallel-storage")
 	cmd.layout = ctx.String("layout")
 	cmd.removeAllRepositories = ctx.StringSlice("remove-all-repositories")
 	cmd.backupID = ctx.String("id")
@@ -52,7 +52,7 @@ func restoreFlags() []cli.Flag {
 		&cli.IntFlag{
 			Name:  "parallel",
 			Usage: "maximum number of parallel backups",
-			Value: int64(runtime.NumCPU()),
+			Value: runtime.NumCPU(),
 		},
 		&cli.IntFlag{
 			Name:  "parallel-storage",
